+++ src/_3_problems_from_21_to_30/_3_5_problem_25/Problem25.cpp
@@ -10,13 +10,13 @@ int readPositiveNumber() {
     return number;
 }
 
-int randomNumber(
-    const int FROM,
-    const int TO
+short randomNumber(
+    const short FROM,
+    const short TO
 ) { return rand() % (TO - FROM + 1) + FROM; }
 
 void fillRandomNumbers(
-    int numbers[],
+    short numbers[],
     int numberCount
 ) {
     while (numberCount--)
@@ -27,7 +27,7 @@ void fillRandomNumbers(
 }
 
 void printNumbers(
-    const int NUMBERS[],
+    const short NUMBERS[],
     const int NUMBER_COUNT
 ) {
     for (int index = 0; index < NUMBER_COUNT; ++index) {
@@ -37,14 +37,14 @@ void printNumbers(
     }
 }
 
-int getMinimumNumber(
-    const int NUMBERS[],
+short getMinimumNumber(
+    const short NUMBERS[],
     int numberCount
 ) {
-    int minimumNumber = NUMBERS[--numberCount];
+    short minimumNumber = NUMBERS[--numberCount];
     while (numberCount--)
         if (
-            const int CURRENT_NUMBER = NUMBERS[numberCount];
+            const short CURRENT_NUMBER = NUMBERS[numberCount];
             CURRENT_NUMBER < minimumNumber
         )
             minimumNumber = CURRENT_NUMBER;
@@ -55,7 +55,7 @@ int main() {
     srand(static_cast<unsigned>(time(nullptr)));
 
     const int NUMBER_COUNT = readPositiveNumber();
-    int numbers[NUMBER_COUNT];
+    short numbers[NUMBER_COUNT];
     fillRandomNumbers(
         numbers,
         NUMBER_COUNT
