+++ src/_3_problems_from_21_to_30/_3_6_problem_26/Problem26.cpp
@@ -10,13 +10,13 @@ int readPositiveNumber() {
     return number;
 }
 
-int randomNumber(
-    const int FROM,
-    const int TO
+short randomNumber(
+    const short FROM,
+    const short TO
 ) { return rand() % (TO - FROM + 1) + FROM; }
 
 void fillRandomNumbers(
-    int numbers[],
+    short numbers[],
     int numberCount
 ) {
     while (numberCount--)
@@ -27,7 +27,7 @@ void fillRandomNumbers(
 }
 
 void printNumbers(
-    const int NUMBERS[],
+    const short NUMBERS[],
     const int NUMBER_COUNT
 ) {
     for (int index = 0; index < NUMBER_COUNT; ++index) {
@@ -38,7 +38,7 @@ void printNumbers(
 }
 
 long numbersSum(
-    const int NUMBERS[],
+    const short NUMBERS[],
     int numberCount
 ) {
     long sum = 0;
@@ -51,7 +51,7 @@ int main() {
     srand(static_cast<unsigned>(time(nullptr)));
 
     const int NUMBER_COUNT = readPositiveNumber();
-    int numbers[NUMBER_COUNT];
+    short numbers[NUMBER_COUNT];
     fillRandomNumbers(
         numbers,
         NUMBER_COUNT
